pyproject.toml
[project]
name = "real-world-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "load-dotenv>=0.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
